mo stupid me smarter than yanni

+++ NumberRecognition/NumberRecognition/ImageFolder/Image.cs
@@ -1,11 +1,11 @@
 ﻿using System.Numerics;
 namespace NumberRecognition.ImageFolder;
 public class Image {
-    public byte[]? ImageData { get; }
+    public float[]? ImageData { get; }
     public float Width { get; }
     public float Height { get; }
     public int Value { get; }
-    public Image(byte[] data, int value, Vector2 ImageScale) {
+    public Image(float[] data, int value, Vector2 ImageScale) {
         ImageData = data;
         Value = value;
         Width = ImageScale.X;
